requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
